rename to senv + build logic
