+++ pyproject.toml
@@ -3,13 +3,13 @@ requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
 
 [tool.poetry]
-name = "pysenv"
+name = "senv"
 version = "0.1.0"
 description = ""
 authors = ["Jorge <jorge.girazabal@gmail.com>"]
 
 [tool.poetry.scripts]
-pysenv = 'pysenv.main:app'
+senv = 'senv.main:app'
 
 [tool.poetry.dependencies]
 python = "^3.7.0"
@@ -30,8 +30,8 @@ pytest-cov = "^2.10.1"
 pytest-xdist = "^2.2.0"
 
 
-[tool.pysenv]
-venv-name = "pysenv_test_name"
+[tool.senv]
+venv-name = "senv_test_name"
 conda-channels = [
     "conda-forge",
     "defaults"
