--- pysenv/errors.py
@@ -1,10 +1,0 @@
-class PysenvError(Exception):
-    pass
-
-
-class PysenvInvalidPythonVersion(PysenvError):
-    pass
-
-
-class PysenvNotSupportedPlatform(PysenvError):
-    pass
