--- pysenv/log.py
@@ -1,5 +1,0 @@
-import logging
-
-
-logging.basicConfig(format="%(message)s", level=logging.INFO)
-log = logging.getLogger(__name__)
