--- pysenv/utils.py
@@ -1,13 +1,0 @@
-import contextlib
-import os
-from pathlib import Path
-
-
-@contextlib.contextmanager
-def cd(path: Path):
-    cwd = os.getcwd()
-    os.chdir(str(path))
-    try:
-        yield
-    finally:
-        os.chdir(cwd)
