+++ senv/errors.py
@@ -1,0 +1,10 @@
+class SenvError(Exception):
+    pass
+
+
+class SenvInvalidPythonVersion(SenvError):
+    pass
+
+
+class SenvNotSupportedPlatform(SenvError):
+    pass
