+++ senv/log.py
@@ -1,0 +1,4 @@
+import logging
+
+logging.basicConfig(format="%(message)s", level=logging.INFO)
+log = logging.getLogger(__name__)
