+++ senv/tests/static/senv_override_pyproject.toml
@@ -1,0 +1,30 @@
+[tool.poetry]
+name = "test_name"
+version = "0.1.0"
+description = ""
+authors = ["author <author@flatiron.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.7.0"
+ensureconda = "^1.4.0"
+condax = "^0.0.5"
+click = "^7.1.2"
+tomlkit = "^0.7.0"
+appdirs = "^1.4.4"
+conda-lock = "^0.7.3"
+
+[tool.poetry.dev-dependencies]
+pyinstaller = "^4.1"
+pytest = "^6.2.1"
+
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
+
+[tool.senv]
+conda-channels = [
+    "conda-forge",
+    "defaults"
+]
+[tool.senv.venv]
+name="overridden_name"
