+++ senv/tests/static/simple_pyproject.toml
