+++ senv/tests/static/small_conda_pyproject.toml
@@ -1,0 +1,15 @@
+[tool.poetry]
+name = "test_name"
+version = "0.1.0"
+description = ""
+authors = ["author <author@flatiron.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.7.0"
+appdirs = "^1.4.4"
+
+[tool.poetry.dev-dependencies]
+pytest = "^6.2.1"
+
+[tool.senv]
+build-system = "conda"
